test: extract tooltip target unit tests to separate file
